Refactored namespaces for tests
