+++ Hexxle.UnitTest/Logic/PointResolverTest.cs
@@ -2,7 +2,7 @@ using Hexxle.Interfaces;
 using Hexxle.TileSystem;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Hexxle.UnitTest
+namespace Hexxle.Tests.Logic
 {
     [TestClass]
     public class PointResolverTest
