+++ unity/Assets/Tests/EditMode/Logic.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 96843c9d5bef12042850d4196670527d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
