+++ unity/Assets/Tests/EditMode/Logic/TileMapTests.cs
@@ -3,7 +3,7 @@ using Hexxle.Logic;
 using Hexxle.TileSystem;
 using Hexxle.CoordinateSystem;
 
-namespace Tests
+namespace Hexxle.Tests.Logic
 {
     public class TileMapTests
     {
